Refactor option parsing in cli and test code
Allows cleaning up header includes, also somewhat smaller binaries.
